refactor: remove components folder
