--- src/pages/Content/common/PageDetect.ts
@@ -1,1 +1,0 @@
-export const isPerceptor = (): boolean => window.location.search.includes('perceptor');
