+++ src/pages/Content/config-service.ts
@@ -1,0 +1,3 @@
+export const loadConfigFromGithub = () => {
+
+}
